Refatoracao do codigo, finalizacao de menus e documentacao

+++ src/Models/Matricula.hs
@@ -9,14 +9,14 @@ import Data.Aeson (ToJSON, FromJSON)
 data Matricula = Matricula
   { _idAluno     :: Int
   , _idTurma     :: Int
-  , _craNoMomento :: Float
+  , _craNoMomento :: Double
   , _status       :: StatusMatricula
   } deriving (Show, Eq, Generic, ToJSON, FromJSON)
 
 data StatusMatricula = Solicitada | Deferida | Indeferida
     deriving (Show, Eq, Generic, ToJSON, FromJSON)
 
-criarMatricula :: Int -> Int -> Float -> Matricula
+criarMatricula :: Int -> Int -> Double -> Matricula
 criarMatricula idA idT cra = Matricula idA idT cra Solicitada
 
 getIdAlunoMatricula :: Matricula -> Int
